fix: use i64 for index file size (#586)

+++ bindings/c/src/tests.rs
@@ -1764,7 +1764,7 @@ fn build_pk_vector_table(path: &str, vectors: &[[f32; PK_DIM]]) -> Table {
         let index_file = IndexFileMeta {
             index_type: INDEX_TYPE.to_string(),
             file_name: index_file_name,
-            file_size: i32::try_from(index_file_size).unwrap(),
+            file_size: i64::try_from(index_file_size).unwrap(),
             row_count: i32::try_from(row_count).unwrap(),
             deletion_vectors_ranges: None,
             global_index_meta: Some(GlobalIndexMeta {
